+++ R/info-git.R
@@ -27,7 +27,7 @@ pkginfo_git_info <- function (path) {
 
         repo <- utils::tail (strsplit (u, "/") [[1]], 1)
         org <- utils::tail (strsplit (u, "/") [[1]], 2) [1]
-        has_token <- length (get_gh_token ()) > 0L
+        has_token <- length (gh::gh_token ()) > 0L
         if (curl::has_internet () & has_token) {
             branch <- get_default_github_branch (org, repo)
         }